Init: iterates dirs, skips already iterated, prints filepaths.
That is: it iterates the configured dirs. Both recursive and
non-recursive iteration is supported. There's also support to skip
certain directories like .hg and .git.
Since recursive iteration is supported it's possible that accidentally
or not the same directory is iterated multiple times. This is prevented
by keeping track of the iterated directories. If one is already iterated
it is skipped but since non-recursive iteration is also supported the
sub directories will still be iterated.
For now just print the files that are n*30 days old.